Move EventDispatcher to watch module
